% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agamma.R
\name{agamma}
\alias{agamma}
\title{Alternative formulation of rgamma function, supplying  mean and shape.}
\usage{
agamma(n, mean, shape)
}
\arguments{
\item{n}{Number of observations, passed to \code{\link[=rgamma]{rgamma()}}}

\item{mean}{mean}

\item{shape}{shape}
}
\description{
Refer to the help for \code{\link[=rgamma]{rgamma()}} for the meaning of \code{shape} and \code{scale}.
}
