% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_resources.R
\name{fast_server_usage_summary}
\alias{fast_server_usage_summary}
\title{Rapidly computes a resource usage summary.}
\usage{
fast_server_usage_summary(.env, name, summarize = FALSE)
}
\arguments{
\item{.env}{A \code{simmer} environment, as defined by \code{\link[=shimmer]{shimmer()}}}

\item{name}{Resource name, e.g. "cpu"}

\item{If}{TRUE, returns the overall mean utilization, i.e. the last row of
compuation}
}
\description{
Rapidly computes a resource usage summary.
}
