% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_resources.R
\name{last_resource_value}
\alias{last_resource_value}
\title{Extracts last available value of a resource}
\usage{
last_resource_value(.env, name)
}
\arguments{
\item{.env}{A \code{simmer} environment, as defined by \code{\link[=shimmer]{shimmer()}}}

\item{name}{Resource name, e.g. "cpu"}
}
\description{
Extracts last available value of a resource
}
\seealso{
Other resource summary functions: 
\code{\link{fast_server_usage_summary}()}
}
\concept{resource summary functions}
