% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer.shiny-package.R
\docType{package}
\name{miniCRAN-package}
\alias{miniCRAN-package}
\alias{_PACKAGE}
\alias{miniCRAN}
\alias{minicran}
\title{simmer.shiny: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Andrie de Vries \email{apdevries@gmail.com} [copyright holder]

}
