% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_shimmer_response_histogram}
\alias{plot_shimmer_response_histogram}
\title{Plot histogram of CPU response times}
\usage{
plot_shimmer_response_histogram(.env, binwidth = 0.1)
}
\arguments{
\item{.env}{A \code{simmer} environment, as defined by \code{\link[=shimmer]{shimmer()}}}

\item{binwidth}{Passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}}
}
\description{
Plot histogram of CPU response times
}
\seealso{
Other plot functions: 
\code{\link{plot_shimmer_connection_usage}()},
\code{\link{plot_shimmer_cpu_usage}()},
\code{\link{plot_shimmer_process_usage}()},
\code{\link{plot_shimmer_rejection_usage}()},
\code{\link{plot_shimmer_resources}()},
\code{\link{plot_shimmer_usage}()}
}
\concept{plot functions}
