% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_shimmer_usage}
\alias{plot_shimmer_usage}
\title{Plot shiny usage}
\usage{
plot_shimmer_usage(.env)
}
\arguments{
\item{.env}{A \code{simmer} environment, as defined by \code{\link[=shimmer]{shimmer()}}}
}
\description{
Plot shiny usage
}
\seealso{
Other plot functions: \code{\link{plot_shimmer_cpu_histogram}},
  \code{\link{plot_shimmer_resources}}
}
