% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_shiny_cpu_histogram}
\alias{plot_shiny_cpu_histogram}
\title{Plot histogram of CPU response times}
\usage{
plot_shiny_cpu_histogram(.env, binwidth = 0.1)
}
\arguments{
\item{.env}{A \code{simmer} environment, as defined by \code{\link[=simmer_shiny]{simmer_shiny()}}}
}
\description{
Plot histogram of CPU response times
}
\seealso{
Other plot functions: \code{\link{plot_shiny_resources}},
  \code{\link{plot_shiny_usage}}
}
