% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_shiny_resources}
\alias{plot_shiny_resources}
\title{Plot shiny resources}
\usage{
plot_shiny_resources(.env)
}
\arguments{
\item{.env}{A \code{simmer} environment, as defined by \code{\link[=simmer_shiny]{simmer_shiny()}}}
}
\description{
Plot shiny resources
}
\seealso{
Other plot functions: \code{\link{plot_shiny_cpu_histogram}},
  \code{\link{plot_shiny_usage}}
}
