% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_shiny_usage}
\alias{plot_shiny_usage}
\title{Plot shiny usage}
\usage{
plot_shiny_usage(.env)
}
\arguments{
\item{.env}{A \code{simmer} environment, as defined by \code{\link[=simmer_shiny]{simmer_shiny()}}}
}
\description{
Plot shiny usage
}
\seealso{
Other plot functions: \code{\link{plot_shiny_cpu_histogram}},
  \code{\link{plot_shiny_resources}}
}
