% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimmer.R
\name{shimmer}
\alias{shimmer}
\title{Defines shiny simulation and runs for a specified period of time.}
\usage{
shimmer(until = 3600, config, config_file)
}
\arguments{
\item{until}{Time in seconds}

\item{config}{Configuration parameters. If not provided, then attempts to read from the \code{config_file}}

\item{config_file}{Path to config file with runtime and app settings.  If neither \code{config} nor \code{config_file} is provided, attempts to read from \code{\link[=shimmer_config_file]{shimmer_config_file()}}.}
}
\value{
a \code{simmer} environment
}
\description{
This uses the \code{simmer} package to set up a simulation run.
}
\seealso{
\code{\link[=shimmer_config_file]{shimmer_config_file()}}
}
