% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimmer.R
\name{shimmer_config_file}
\alias{shimmer_config_file}
\title{Defines the location of the example config.yml}
\usage{
shimmer_config_file()
}
\description{
Defines the location of the example config.yml
}
\seealso{
\code{\link[=simmer_shiny]{simmer_shiny()}}
}
