% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimmer_shiny.R
\name{shimmer_shiny}
\alias{shimmer_shiny}
\title{Runs a shimmer simulation inside a shiny app.}
\usage{
shimmer_shiny()
}
\description{
Runs a shimmer simulation inside a shiny app.
}
\seealso{
\code{\link[=shimmer]{shimmer()}}
}
