% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer_shiny.R
\name{simmer_shiny}
\alias{simmer_shiny}
\title{Defines shiny simulation and runs for a specified period of time}
\usage{
simmer_shiny(until = 3600, config_file = system.file("R/config.yml", package
  = "simmer.shiny"))
}
\arguments{
\item{until}{Time in seconds}

\item{config_file}{Path to config file with runtime and app settings}
}
\value{
a \code{simmer} environment
}
\description{
Defines shiny simulation and runs for a specified period of time
}
