% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_resources.R
\name{summary_cpu}
\alias{summary_cpu}
\title{Extract summary value}
\usage{
summary_cpu(.env)
}
\arguments{
\item{.env}{A \code{simmer} environment, as defined by \code{\link[=shimmer]{shimmer()}}}
}
\description{
Extract summary value
}
\seealso{
Other plot functions: \code{\link{plot_shimmer_connection_usage}},
  \code{\link{plot_shimmer_cpu_usage}},
  \code{\link{plot_shimmer_process_usage}},
  \code{\link{plot_shimmer_rejection_usage}},
  \code{\link{plot_shimmer_resources}},
  \code{\link{plot_shimmer_response_histogram}},
  \code{\link{plot_shimmer_usage}}
}
