% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_shimmer_rejection_usage}
\alias{plot_shimmer_rejection_usage}
\title{Plot shimmer rejections}
\usage{
plot_shimmer_rejection_usage(.env)
}
\arguments{
\item{.env}{A \code{simmer} environment, as defined by \code{\link[=shimmer]{shimmer()}}}
}
\description{
Plot shimmer rejections
}
\seealso{
Other plot functions: \code{\link{plot_shimmer_connection_usage}},
  \code{\link{plot_shimmer_cpu_usage}},
  \code{\link{plot_shimmer_process_usage}},
  \code{\link{plot_shimmer_resources}},
  \code{\link{plot_shimmer_response_histogram}},
  \code{\link{plot_shimmer_usage}}
}
