% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_shimmer_resources}
\alias{plot_shimmer_resources}
\title{Plot shiny resources}
\usage{
plot_shimmer_resources(.env)
}
\arguments{
\item{.env}{A \code{simmer} environment, as defined by \code{\link[=shimmer]{shimmer()}}}
}
\description{
Plot shiny resources
}
\seealso{
Other plot functions: \code{\link{plot_shimmer_cpu_histogram}},
  \code{\link{plot_shimmer_usage}}
}
