% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimmer-package.R
\docType{package}
\name{shimmer-package}
\alias{shimmer-package}
\title{Discrete Event Simulation for Scaling Shiny, Using the Simmer Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Discrete Event Simulation for Shiny, Using the Simmer Package.
}
\author{
\strong{Maintainer}: Andrie de Vries \email{apdevries@gmail.com} [copyright holder]

}
