% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimmer-package.R
\docType{package}
\name{shimmer-package}
\alias{shimmer-package}
\title{Discrete Event Simulation for Scaling Shiny, Using the Simmer Package}
\description{
Discrete Event Simulation for Scaling Shiny, Using the Simmer Package
}
